requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.28
pymongo>=4.0
PyMySQL>=1.0
oracledb>=2.0
